requirements.txt
# FastAPI Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3
itsdangerous>=2.1.2

# Database
supabase>=2.3.0

# Scraping
playwright>=1.48.0
httpx>=0.26.0
beautifulsoup4>=4.12.0

# Utilities
pandas>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
